Version 0.5.0

+++ src/main/kotlin/com/estonianport/agendaza/dto/PagoDTO.kt
@@ -3,7 +3,7 @@ package com.estonianport.agendaza.dto
 import com.estonianport.agendaza.model.MedioDePago
 import java.time.LocalDateTime
 
-class PagoDTO(val id: Long, val monto : Int, val codigo : String, val medioDePago: MedioDePago,
+class PagoDTO(val id: Long, val monto : Double, val codigo : String, val medioDePago: MedioDePago,
               val nombreEvento: String, val fecha: LocalDateTime) {}
 
 class CodigoEmpresaId(val codigo : String, val empresaId: Long) {}
